Fix sign-conversion and conversion warnings in headers and Makefile
- cutils.h: fix max_uint/min_uint return types (int -> unsigned int)
- util.h: add casts for char32_t/size_t/unsigned char conversions; fix
  qe_iswlower/qe_iswupper return types (char32_t -> int)
- qe.h: add casts in eb_peekc and cp_set_style/cp_set_style1
- color.h: add cast in color_y to silence signed overflow warning
- Makefile: suppress sign-conversion/conversion warnings for vendored
  code (libunicode.c, libregexp.c, lua-amalg.c)

+++ color.h
@@ -195,7 +195,7 @@ int css_get_font_family(const char *str);
 int css_get_enum(const char *str, const char *enum_str);
 int color_dist(QEColor c1, QEColor c2);
 static inline int color_y(QEColor c) {
-    return 30 * ((c >> 16) & 0xff) + 59 * ((c >> 8) & 0xff) + 11 * (c & 0xff);
+    return (int)(30 * ((c >> 16) & 0xff) + 59 * ((c >> 8) & 0xff) + 11 * (c & 0xff));
 }
 
 typedef struct CSSRect {
